APPS-1121 Update integration test suite (feat) ExpectedFailure.py - implementation of failing tests (feat) RegisteredTestFactory.py - refactored, moved to a separate module
