+++ dxcint/config/medium.json
@@ -1,1 +1,1 @@
-{"expected_failure": ["just_fail_wf", "bad_status", "bad_status2"]}
+{"expected_failure": ["bad_status2", "bad_status", "just_fail_wf"]}
