+++ dxcint/dxcint/RegisteredTest.py
@@ -17,29 +17,6 @@ class RegisteredTestError(Exception):
     """
 
 
-class RegisteredTestFactory(object):
-    @classmethod
-    def register_test(
-            cls,
-            src_file: str,
-            category: str,
-            test_name: str,
-            context: Context
-    ):
-        test_type_switch = {
-            "mock_category": RegisteredTest,        # For testing only
-            # ADD NEW CATEGORY HERE
-        }
-        registered_test = test_type_switch.get(category, None)
-        if not registered_test:
-            raise RegisteredTestError(
-                f"RegisteredTestFactory.register_test(): Category {category} is not recognized. Existing categories "
-                f"are {test_type_switch.keys()}. Add new category and new test implementation as a subclass of "
-                f"`RegisteredTest`."
-            )
-        return registered_test(src_file, category, test_name, context)
-
-
 class RegisteredTest(object):
     def __init__(
             self,
@@ -192,7 +169,6 @@ class RegisteredTest(object):
             "passed": False,
             "message": "This is an abstract base class. Use a concrete implementation"
         }
-
 
     def _create_messenger(self) -> Messenger:
         raise RegisteredTestError("Messenger is not implemented")
